[GST-1327] Adds aws cluster discovery to Kafka
This allows kafka machines to discover which id and hostname they
should use in AWS, this will allow machines to be replaced if they die
without disruption/manual intervention. Rolling AMI updates are also
possible.

Also added id and advertised hostname settings so a machine can be
recognised properly in a cluster, Kafka will auto rebalance data if these
settings are consistent, so if a machine dies or is replace it will
automatically sync data from the rest of the cluster if the hostname
and id are recognised.