[GITHUB-13] Allows broker id to be set via tag
Adds ability to specify broker id via a tag upfront, this
removes the need for the aws cluster autodiscovery script
when using this method.

The idea is to use a single ASG per instance, the ASG specifies
the broker.id via a tag which is picked up the instance, in
addition an EBS volume and ENI keep the instances data and
IP address constant (hence the addition of the interface
options).

Other changes:

* Upgrade to verson 0.10.1 by default
* Adds ability to specify an interface for bind and/or advertise
* Updates Kafka listener config to latest format